Ai_project/APIServer/requirements.txt
fastapi
uvicorn
transformers
torch
sentencepiece
protobuf
openai-whisper
python-multipart
